disable host image copy to make it work on windows with RTX1070

+++ vulkan/VulkanImage.hpp
@@ -72,7 +72,7 @@ class VulkanImage : public VulkanResource {
     int getImageChannelSize() const { return m_chansize_; }
     int getImageChannelNum() const { return m_chans_; }
 
-    void splitImageView(std::vector<int64_t> &layers);
+    void splitImageView(std::vector<int> &layers);
 
   private:
     VkExtent3D m_dim_;
